backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
bcrypt==4.3.0
PyMySQL==1.1.1
PyJWT==2.7.0
requests==2.31.0
python-dotenv==1.1.0
